[Gold IV] Title: 최단경로, Time: 312 ms, Memory: 15288 KB -BaekjoonHub

+++ 백준/Gold/1753. 최단경로/최단경로.cc
@@ -25,8 +25,8 @@ int main() {
 		pq.pop();
 		if (dist[here] != here_dist) continue;
 		for (pairll there : adj[here]) {
-			int _dist = there.first;
-			int _there = there.second;
+			ll _dist = there.first;
+			ll _there = there.second;
 			if (dist[_there] > dist[here] + _dist) {
 				dist[_there] = dist[here] + _dist;
 				pq.push({ dist[_there], _there });
@@ -40,3 +40,4 @@ int main() {
 	}
 	return 0;
 }
+
